(#210) Introduce linizer
